Finished up the functions for scoreboard
Used DOM manipulation, added wins, losses, ties.

Next goals (No specific order):
Need to clean up scoreboard with CSS
Also change the gifs to static images until user hovers over them
Allow NPC choice to fade in/out and only play once it's called for.
Add a h2 element above the game that mentions which round it is currently on.
Create a Play button that takes user input for rounds.
Change the scoreboard so that it can tally the rounds and add the score under the players.