Made more changes
Added the h2 element above the game, it now mentions the current round and set rounds

Next goals:
Change the scoreboard so that it can tally the rounds
Make an input for the user that'll reveal itself when pressing Play now button
Change "Pick one" to the player option they picked.

Low Priority:
Adding animation for NPC when it plays
Add a delay so it seems the NPC is thinking.
Add a win/loss/tie response from the NPC when playing against it.